==> udacity_data_science/7_data_engineering/1_etl_pipeline/3_sql_exercise.ipynb <==
# Extract Data from SQL Databases

In this workbook, you'll gain experience extracting data from SQL databases. This is an overview of Python tools and assumes you already have experience writing SQL queries.

# Pandas and sqlite3

You can use Pandas to open a SQL database or to run a SQL query against a database. There is more than one way to do this depending on the type of SQL database you are working with: the [sqlite3 library](https://www.sqlite.org/about.html) or the [sqlalchemy library](https://www.sqlalchemy.org/). The sqlite3 library can handle sqlite databases. SQLAlchemy can handle sqlite3 as well as other types of SQL databases like MySQL and PostgreSQL.

In the same folder as this Jupyter notebook, there is a SQLite database file called "population_data.db". SQLite is a database engine meant for single applications. The entire database is contained in one file. You can read more about SQLite [here](https://www.sqlite.org/whentouse.html). 

In this example, the "population_data.db" database contains only one table called "population_data". Run the code in the following cells to see how to use a SQLite database with pandas. (If you're curious how the data was converted from a csv file to a database, go to File->Open and click on create_db.py).

# SQLite3 and Pandas Example

import sqlite3
import pandas as pd

# connect to the database
conn = sqlite3.connect('population_data.db')

# run a query
pd.read_sql('SELECT * FROM population_data', conn)

pd.read_sql('SELECT "Country_Name", "Country_Code", "1960" FROM population_data', conn)

# SQLAlchemy and Pandas example

If you are working with a different type of database such as MySQL or PostgreSQL, you can use the SQLAlchemy library with pandas. Here are the instructions for connecting to [different types of databases using SQLAlchemy](http://docs.sqlalchemy.org/en/latest/core/engines.html).

Run the code below to see how to connect to the population_data.db database.

import pandas as pd
from sqlalchemy import create_engine

### 
# create a database engine 
# to find the correct file path, use the python os library:
import os
print(os.getcwd())
#
###

engine = create_engine('sqlite:////Users/tobias/Code/_data/DSND_Term2/lessons/ETLPipelines/3_sql_exercise/population_data.db')
pd.read_sql("SELECT * FROM population_data", engine)

# Exercise

Connect to the population_data.db SQLite database, and answer the following questions:
1. Write a query that finds the change in population from 1960 to 1961 in Aruba
2. Write a query that finds the population of Belgium and also Luxembourg in 1975. The output should have two rows.

There is a solution if you go to File->Open->3_sql_exercise_solution.ipynb

pd.read_sql('SELECT "1960" - "1961" FROM population_data WHERE Country_Name IS "Aruba"', engine)

pd.read_sql('SELECT Country_Name, "1975" FROM population_data WHERE Country_Name IS "Belgium" OR Country_Name IS "Luxembourg"', engine)
