==> udacity_data_science/7_data_engineering/1_etl_pipeline/5_combining_data.ipynb <==
# Combining Data

Practice combining data from two different data sets. In the same folder as this Jupyter notebook, there are two csv files:
* rural_population_percent.csv
* electricity_access_percent.csv

They both come from the World Bank Indicators data. 
* https://data.worldbank.org/indicator/SP.RUR.TOTL.ZS
* https://data.worldbank.org/indicator/EG.ELC.ACCS.ZS

The rural populaton data represents the percent of a country's population that is rural over time. The electricity access data shows the percentage of people with access to electricity.

In this exercise, you will combine these two data sets together into one pandas data frame.

# Exercise 1

Combine the two data sets using the [pandas concat method](https://pandas.pydata.org/pandas-docs/stable/merging.html). In other words, find the union of the two data sets.

# TODO: import the pandas library
import pandas as pd

# TODO: read in each csv file into a separate variable
# HINT: remember from the Extract material that these csv file have some formatting issues
# HINT: The file paths are 'rural_population_percent.csv' and 'electricity_access_percent.csv'
df_rural = pd.read_csv("rural_population_percent.csv", skiprows=3)
df_electricity = pd.read_csv("electricity_access_percent.csv", skiprows=3)

# TODO: remove the 'Unnamed:62' column from each data set
df_rural = df_rural.drop(columns=["Unnamed: 62"])
df_electricity = df_electricity.drop(columns=["Unnamed: 62"])

# TODO: combine the two data sets together using the concat method
df = pd.concat([df_rural, df_electricity])
df.head()

# Exercise 2 (Challenge)

This exercise is more challenging.

The resulting data frame should look like this:

|Country Name|Country Code|Year|Rural_Value|Electricity_Value|
|--|--|--|--|--|--|
|Aruba|ABW|1960|49.224|49.239|
... etc.

Order the results in the dataframe by country and then by year

Here are a few pandas methods that should be helpful:
* [melt](http://pandas.pydata.org/pandas-docs/stable/generated/pandas.melt.html)
* [drop](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.drop.html)
* [merge](https://pandas.pydata.org/pandas-docs/version/0.23/generated/pandas.DataFrame.merge.html)
* [sort_values](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.sort_values.html)

HINT: You can use country name, country code, and the year as common keys between the data sets

# TODO: merge the data sets together according to the instructions. First, use the 
# melt method to change the formatting of each data frame so that it looks like this:
# Country Name, Country Code, Year, Rural Value
# Country Name, Country Code, Year, Electricity Value
year_columns = df_rural.columns[4:].to_list()

df_rural_melted = df_rural.melt(id_vars=["Country Name", "Country Code"], value_vars=year_columns)
df_electricity_melted = df_electricity.melt(id_vars=["Country Name", "Country Code"], value_vars=year_columns)

# TODO: drop any columns from the data frames that aren't needed
# ?
df_rural_melted.rename(columns={'variable': 'Year', 'value': 'Rural Value'}, inplace=True)
df_electricity_melted.rename(columns={'variable': 'Year', 'value': 'Electricity Value'}, inplace=True)

# TODO: merge the data frames together based on their common columns
# in this case, the common columns are Country Name, Country Code, and Year
df = pd.merge(df_rural_melted, df_electricity_melted, how="outer", on=["Country Name", "Country Code", "Year"])

# TODO: sort the results by country and then by year

df_combined = df.sort_values(["Country Name", "Year"])
df_combined
